requirements.txt
# TickEye 基金监控项目依赖

# Python 版本要求: >= 3.9 (AKShare 要求)

# 核心数据获取
akshare>=1.12.0          # 基金数据获取 (金融数据接口)
requests>=2.28.0         # HTTP客户端 (AKShare 依赖)

# 数据处理和分析
pandas>=1.5.0            # 数据处理和分析 (核心依赖)
numpy>=1.21.0            # 数值计算 (pandas 依赖)

# 任务调度 (可选，未来扩展用)
schedule>=1.2.0          # 简单任务调度
apscheduler>=3.10.0      # 高级任务调度 (可选)

# 配置文件处理 (未来扩展用)
PyYAML>=6.0              # YAML配置文件解析
python-dotenv>=1.0.0     # .env 文件环境变量加载

# 日期和时间处理
python-dateutil>=2.8.0   # 日期处理工具
pytz>=2022.1             # 时区处理

# 网络和HTTP (AKShare 的依赖)
urllib3>=1.26.0          # HTTP 库
certifi>=2022.12.7       # SSL 证书验证

# 开发和测试工具
pytest>=7.0.0            # 单元测试框架 (开发环境)
pytest-cov>=4.0.0        # 测试覆盖率 (开发环境)

# 代码质量工具 (开发环境)
flake8>=5.0.0            # 代码风格检查
black>=22.0.0            # 代码格式化
isort>=5.12.0            # import 排序

# 类型检查 (开发环境，可选)
mypy>=1.0.0              # 静态类型检查

# 数据验证 (未来扩展用，可选)
# pydantic>=1.10.0       # 数据验证和序列化

# 通知扩展 (未来功能，可选)
# requests-oauthlib>=1.3.0  # OAuth 认证 (飞书机器人等)

# Web界面 (未来扩展，可选)
# flask>=2.2.0           # Web框架
# jinja2>=3.1.0          # 模板引擎
# gunicorn>=20.1.0       # WSGI 服务器

# 性能监控 (可选)
# psutil>=5.9.0          # 系统性能监控

# 安装说明:
# 1. 基础安装: pip install -r requirements.txt
# 2. 开发环境: pip install -r requirements.txt pytest flake8 black mypy
# 3. 中国用户可使用: pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/